+++ bld/wv/c/dlgexpr.c
@@ -48,7 +48,7 @@
 
 
 typedef bool rtn_func(const char *,void *);
-typedef bool dlg_func(const char *,char *,unsigned);
+typedef bool dlg_func(const char *,char *,size_t);
 
 /* to be moved to header files ! */
 extern char             *StrDouble(xreal*,char*);
@@ -94,7 +94,7 @@ void DlgSetLong( gui_window *gui, gui_ctl_id id, long value )
     GUISetText( gui, id, TxtBuff );
 }
 
-static bool     DlgGetItemWithRtn( char *buff, unsigned buff_len, const char *title,
+static bool     DlgGetItemWithRtn( char *buff, size_t buff_len, const char *title,
                                    void *value, rtn_func *rtn, dlg_func *dlg )
 {
     bool        rc;
@@ -110,7 +110,7 @@ static bool     DlgGetItemWithRtn( char *buff, unsigned buff_len, const char *ti
     }
 }
 
-static bool     DlgGetItem( char *buff, unsigned buff_len, const char *title, void *value, rtn_func *rtn )
+static bool     DlgGetItem( char *buff, size_t buff_len, const char *title, void *value, rtn_func *rtn )
 {
     return( DlgGetItemWithRtn( buff, buff_len, title, value, rtn, DlgNewWithSym ) );
 }
@@ -123,7 +123,7 @@ bool    DlgLongExpr( const char *title, long *value )
     return( DlgGetItem( new, EXPR_LEN, title, value, (rtn_func *)DlgScanLong ) );
 }
 
-bool    DlgAnyExpr( const char *title, char *buff, unsigned buff_len )
+bool    DlgAnyExpr( const char *title, char *buff, size_t buff_len )
 {
     return( DlgGetItem( buff, buff_len, title, NULL, (rtn_func *)DlgScanAny ) );
 }
