+++ bld/wv/wdb/c/wdbstub.c
@@ -189,7 +189,7 @@ bool ProcessCmdHelp( char *param );
 bool IsCmdTest( const char *cmd );
 bool ProcessCmdTest( const char *param );
 bool ProcessDebuggerCmd( char *cmd );
-bool DlgNewWithSym( const char *title, char *buff, int buff_len );
+bool DlgNewWithSym( const char *title, char *buff, size_t buff_len );
 bool DlgUpTheStack( void );
 bool DlgAreYouNuts( unsigned long mult );
 bool DlgBackInTime( bool warn );
@@ -1572,7 +1572,7 @@ bool DUIDlgGivenAddr( const char *title, address *value )
     // needed when segment's don't map (from new/sym command)
     return( false );
 }
-bool DlgNewWithSym( const char *title, char *buff, int buff_len )
+bool DlgNewWithSym( const char *title, char *buff, size_t buff_len )
 {
     // used by print command with no arguments
     return( true );
