+++ bld/wv/win/c/stubdb.c
@@ -73,7 +73,7 @@
 /*********************************************/
 /* TODO! review all these prototypes declaration if they are local(static) or external */
 
-bool DlgNewWithSym( const char *title, char *buff, int buff_len );
+bool DlgNewWithSym( const char *title, char *buff, size_t buff_len );
 bool DlgUpTheStack( void );
 bool DlgAreYouNuts( unsigned long mult );
 bool DlgBackInTime( bool warn );
@@ -346,7 +346,7 @@ bool DUIDlgGivenAddr( const char *title, address *value )
     // needed when segment's don't map (from new/sym command)
     return( false );
 }
-bool DlgNewWithSym( const char *title, char *buff, int buff_len )
+bool DlgNewWithSym( const char *title, char *buff, size_t buff_len )
 {
     // used by print command with no arguments
     return( true );
